pyproject.toml
[project]
name = "esfs"
description = "esfs"
version = "0.0.1"
authors = [
    { name = "Arthur Radley", email = "arthur_radley@hotmail.co.uk"}
]
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.11"
dependencies = [
    "anndata==0.11.4",
    "numpy==1.26.4",
    "pandas==2.2.3",
    "scipy==1.11.4",
    "matplotlib==3.7.3",
    "scanpy[leiden]==1.11.0",
    "scikit-learn==1.5.2",
    "plotly[express]==6.5.1",
    "multiprocess==0.70.17",
    "p-tqdm==1.4.2",
    "umap-learn==0.5.7",
    "ipython==8.12.3",
    "zarr<3"
]
classifiers=[
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
gpu = [
    "cupy-cuda12x"
]
mlx = [
    "mlx>=0.20.0"
]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]